third_party/blink/web_tests/external/wpt/css/css-values/attr-security-expected.txt

This is a testharness.js-based test.
[FAIL] '--x: attr(data-foo type(<url>))' with data-foo="url(https://does-not-exist.test/404.png)"
  assert_equals: expected "url(\\"https://does-not-exist.test/404.png\\")" but got ""
Harness: the test ran to completion.
